+++ package.json
@@ -61,6 +61,7 @@
     "@types/react": "^18.0.21",
     "@types/react-dom": "^18.0.6",
     "@types/react-helmet": "^6.1.5",
+    "@types/react-scroll": "^1.8.4",
     "babel-eslint": "^10.1.0",
     "eslint": "^7.14.0",
     "eslint-config-airbnb": "^18.2.1",
