+++ src/context/context.ts
@@ -1,6 +1,7 @@
 import React from 'react';
+import type { PortfolioContextType } from '../types';
 
-const PortfolioContext = React.createContext();
+const PortfolioContext = React.createContext<PortfolioContextType>({});
 
 export const PortfolioProvider = PortfolioContext.Provider;
 export const PortfolioConsumer = PortfolioContext.Consumer;
