+++ src/images/index.d.ts
@@ -1,0 +1,3 @@
+declare module '*.jpg';
+declare module '*.png';
+declare module '*.svg';
