--- src/mock/aboutMeText.js
@@ -1,8 +1,0 @@
-export const paragraphOne =
-  "I'm a software engineer in the Bay Area with broad experience working with JavaScript/React frontends and several backend setups, including Ruby on Rails, Node/ExpressJS, Google Firebase, and AWS AppSync. I love working on all aspects of an app, but I'm especially drawn to creating new components/pages on the frontend and working with business logic on the backend.";
-
-export const paragraphTwo =
-  "Before I got into software development, I was a project manager in the electric utility industry, where I coordinated 3 full crews' activities. I've always built my own tools to solve specific problems (usually Excel calculators), and I took that a step further when I founded my own productivity app aimed at people working in the same position. In addition to honing my technical skills, this experience has taught me about making significant UI/UX decisions, locking down information security, and making apps accessible to everyone.";
-
-export const paragraphThree =
-  "Outside of work, I enjoy traveling in the off-season, cycling along the Bay Trail, and tinkering with projects at ACE Makerspace in Oakland. I'm proudest of some of my woodworking projects, including a set of maple wood bookshelf speakers and my dining room table.";
